refactor: util.py を editor.py に rename
